+++ src/js/core/policies.js
@@ -158,6 +158,16 @@ const policies = {
     type : 'boolean'
   },
 
+  DisableSafeMode : {
+    description : browser.i18n.getMessage('policy_description_DisableSafeMode'),
+    enterprise_only : false,
+    first_available : { mainstream : '60.0', esr : '60.0' },
+    info_link : null,
+    is_lockable : false,
+    ui_category : 'disable-features',
+    type : 'boolean'
+  },
+
   DisableProfileRefresh : {
     description : browser.i18n.getMessage('policy_description_DisableProfileRefresh'),
     enterprise_only : false,