+++ CHANGELOG.md
@@ -26,6 +26,8 @@
   [#90](https://github.com/cadeyrn/enterprise-policy-generator/issues/90)
 - policy to add a custom support menu item to the help menu is also available in Firefox ESR, see
   [#90](https://github.com/cadeyrn/enterprise-policy-generator/issues/90)
+- re-added DisableSafeMode policy (removed in Enterprise Policy Generator 4.1.0) but with better description, see
+  [#94](https://github.com/cadeyrn/enterprise-policy-generator/issues/94)
 - improved labels for Certificates|Add policy, see
   [#95](https://github.com/cadeyrn/enterprise-policy-generator/issues/95)
 - allow data:image URI for IconURL property of SearchEngine|Add policy, see